+++ graphistry/layout/graph/graph.py
@@ -8,31 +8,55 @@ import pandas as np
 
 
 class Graph(object):
-    """
-        The graph is stored in disjoint-sets holding each connected component in `components` as a list of graph_core objects.
+    # """
+    #     The graph is stored in disjoint-sets holding each connected component in `components` as a list of graph_core objects.
 
-        **Attributes**
-            C (list[GraphBase]): list of graph_core components.
+    #     **Attributes**
+    #         C (list[GraphBase]): list of graph_core components.
 
-        **Methods**
-            add_vertex(v): add vertex v into the Graph as a new component
-            add_edge(e): add edge e and its vertices into the Graph possibly merging the
-            associated graph_core components
-            get_vertices_count(): see order()
-            vertices(): see graph_core
-            edges(): see graph_core
-            remove_edge(e): remove edge e possibly spawning two new cores
-            if the graph_core that contained e gets disconnected.
-            remove_vertex(v): remove vertex v and all its edges.
-            order(): the order of the graph (number of vertices)
-            norm(): the norm of the graph (number of edges)
-            deg_min(): the minimum degree of vertices
-            deg_max(): the maximum degree of vertices
-            deg_avg(): the average degree of vertices
-            eps(): the graph epsilon value (norm/order), average number of edges per vertex.
-            connected(): returns True if the graph is connected (i.e. it has only one component).
-            components(): returns the list of components
-    """
+
+    #     **add_edge(e):**
+    #         add edge e and its vertices into the Graph possibly merging the associated graph_core components
+
+    #     **get_vertices_count():** 
+    #         see order()
+
+    #     **vertices():** 
+    #         see graph_core
+        
+    #     **edges():** 
+    #         see graph_core
+
+    #     **remove_edge(e):** 
+    #         remove edge e possibly spawning two new cores if the graph_core that contained e gets disconnected.
+
+    #     **remove_vertex(v):** 
+    #         remove vertex v and all its edges.
+
+    #     **order():** 
+    #         the order of the graph (number of vertices)
+
+    #     **norm():** 
+    #         the norm of the graph (number of edges)
+
+    #     **deg_min():** 
+    #         the minimum degree of vertices
+        
+    #     **deg_max():** 
+    #         the maximum degree of vertices
+
+    #     **deg_avg():** 
+    #         the average degree of vertices
+
+    #     **eps():** 
+    #         the graph epsilon value (norm/order), average number of edges per vertex.
+
+    #     **connected():** 
+    #         returns True if the graph is connected (i.e. it has only one component).
+
+    #     **components():**
+    #         returns the list of components
+    # """
 
     component_class = GraphBase
 
@@ -73,16 +97,22 @@ class Graph(object):
             self.components.append(self.component_class(vertices, edge_set, directed))
 
     def add_vertex(self, v):
+        """
+        add vertex v into the Graph as a new component
+        """
         for c in self.components:
             if v in c.verticesPoset:
                 return c.verticesPoset.get(v)
         g = self.component_class(directed = self.directed)
         v = g.add_single_vertex(v)
         self.components.append(g)
+        print("add vertex v into the Graph as a new component")
         return v
 
     def add_edge(self, e):
-
+        """
+        add edge e and its vertices into the Graph possibly merging the associated graph_core components
+        """
         x = e.v[0]
         y = e.v[1]
         x = self.add_vertex(x)
@@ -116,6 +146,9 @@ class Graph(object):
         return None
 
     def vertices(self):
+        """
+         see graph_core
+        """
         for c in self.components:
             vertices = c.verticesPoset
             for v in vertices:
@@ -128,6 +161,9 @@ class Graph(object):
                 yield e
 
     def remove_edge(self, e):
+        """
+        remove edge e possibly spawning two new cores if the graph_core that contained e gets disconnected.
+        """
         # get the GraphBase:
         c = e.v[0].component
         assert c == e.v[1].component
@@ -147,6 +183,9 @@ class Graph(object):
         return e
 
     def remove_vertex(self, x):
+        """
+        remove vertex v and all its edges.
+        """
         c = x.component
         if c not in self.components:
             return None
@@ -165,24 +204,42 @@ class Graph(object):
         return x
 
     def order(self):
+        """
+        the order of the graph (number of vertices)
+        """
         return sum([c.order() for c in self.components])
 
     def norm(self):
+        """
+        the norm of the graph (number of edges)
+        """
         return sum([c.norm() for c in self.components])
 
     def deg_min(self):
+        """
+         the minimum degree of vertices
+        """
         return min([c.deg_min() for c in self.components])
 
     def deg_max(self):
+        """
+         the maximum degree of vertices
+        """
         return max([c.deg_max() for c in self.components])
 
     def deg_avg(self):
+        """
+        the average degree of vertices
+        """
         t = 0.0
         for c in self.components:
             t += sum([v.degree() for v in c.verticesPoset])
         return t / float(self.order())
 
     def eps(self):
+        """
+        the graph epsilon value (norm/order), average number of edges per vertex.
+        """
         return float(self.norm()) / self.order()
 
     def path(self, x, y, f_io = 0, hook = None):
@@ -203,4 +260,7 @@ class Graph(object):
         return r
 
     def connected(self):
+        """
+        returns the list of components
+        """
         return len(self.components) == 1