+++ graphistry/plugins/cugraph.py
@@ -239,16 +239,19 @@ def compute_cugraph(
     
     **Example: Pagerank**
         ::
+
             g2 = g.compute_cugraph('pagerank')
             assert 'pagerank' in g2._nodes.columns
 
     **Example: Katz centrality with rename**
         ::
+
             g2 = g.compute_cugraph('katz_centrality', out_col='katz_centrality_renamed')
             assert 'katz_centrality_renamed' in g2._nodes.columns
 
     **Example: Pass params to cugraph**
         ::
+        
             g2 = g.compute_cugraph('k_truss', params={'k': 2})
             assert 'k_truss' in g2._nodes.columns
 
@@ -360,6 +363,7 @@ def layout_cugraph(
 
     **Example: ForceAtlas2 layout**
         ::
+
             import graphistry, pandas as pd
             edges = pd.DataFrame({'s': ['a','b','c','d'], 'd': ['b','c','d','e']})
             g = graphistry.edges(edges, 's', 'd')
@@ -367,6 +371,7 @@ def layout_cugraph(
 
     **Example: Change which column names are generated**
         ::
+
             import graphistry, pandas as pd
             edges = pd.DataFrame({'s': ['a','b','c','d'], 'd': ['b','c','d','e']})
             g = graphistry.edges(edges, 's', 'd')
@@ -377,6 +382,7 @@ def layout_cugraph(
 
     **Example: Pass parameters to layout methods**
         ::
+        
             import graphistry, pandas as pd
             edges = pd.DataFrame({'s': ['a','b','c','d'], 'd': ['b','c','d','e']})
             g = graphistry.edges(edges, 's', 'd')